refactor: Add unit tests for library routes, scan helpers, and navigation widgets
