+++ tests/test_scan_helpers.py
@@ -1,0 +1,139 @@
+from __future__ import annotations
+
+import sqlite3
+import tempfile
+import time
+import unittest
+from pathlib import Path
+from unittest.mock import patch
+
+from tests.test_support import make_fs_track, touch_text
+from db.database import add_tracks, get_library_file_index, initialize_database
+from library.scan_library import (
+    MutagenError,
+    get_audio_file_signature,
+    iter_audio_paths,
+    new_fs_track_from_path,
+    preview_audio_path_exclusions,
+)
+from ui.workers.library_scanner import LibraryScanner
+from core.models import FsTrack
+
+
+class ScanLibraryHelpersTests(unittest.TestCase):
+    def test_get_audio_file_signature_includes_sidecars(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            root = Path(tmp)
+            audio = root / "track.mp3"
+            txt = root / "track.txt"
+            lrc = root / "track.lrc"
+
+            touch_text(audio, "audio")
+            time.sleep(0.02)
+            touch_text(txt, "plain lyrics")
+            time.sleep(0.02)
+            touch_text(lrc, "[00:01.00]synced")
+
+            sig = get_audio_file_signature(str(audio))
+            self.assertIsNotNone(sig[0])
+            self.assertEqual(sig[1], audio.stat().st_size + txt.stat().st_size + lrc.stat().st_size)
+            self.assertEqual(sig[0], max(audio.stat().st_mtime, txt.stat().st_mtime, lrc.stat().st_mtime))
+
+    def test_iter_audio_paths_and_preview_apply_path_and_regex_exclusions(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            root = Path(tmp)
+            include_a = root / "Music" / "keep.mp3"
+            include_b = root / "Music" / "sub" / "keep.flac"
+            excluded_dir_file = root / "Podcasts" / "skip.mp3"
+            excluded_regex_file = root / "Music" / "demo_track.mp3"
+            non_audio = root / "Music" / "note.txt"
+
+            for path in (include_a, include_b, excluded_dir_file, excluded_regex_file, non_audio):
+                touch_text(path, "x")
+
+            paths = iter_audio_paths(
+                [str(root)],
+                excluded_paths=str(root / "Podcasts"),
+                excluded_patterns=r"demo",
+            )
+            self.assertEqual({Path(p).name for p in paths}, {"keep.mp3", "keep.flac"})
+
+            included, excluded = preview_audio_path_exclusions(
+                [str(root)],
+                excluded_paths=str(root / "Podcasts"),
+                excluded_patterns=r"demo",
+            )
+            self.assertEqual({Path(p).name for p in included}, {"keep.mp3", "keep.flac"})
+            self.assertEqual({Path(p).name for p in excluded}, {"demo_track.mp3"})
+
+    def test_new_fs_track_from_path_skips_corrupt_files(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            audio = Path(tmp) / "broken.mp3"
+            touch_text(audio, "not really audio")
+
+            with patch("library.scan_library.MutagenFile", side_effect=MutagenError("bad frame sync")):
+                track = new_fs_track_from_path(str(audio))
+
+            self.assertIsNone(track)
+
+
+class LibraryScannerIncrementalTests(unittest.TestCase):
+    def test_library_scanner_skips_unchanged_reindexes_new_and_removes_missing(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            db = initialize_database(tmp)
+            db_path = str(Path(tmp) / "db.sqlite3")
+            music_dir = Path(tmp) / "Music"
+            music_dir.mkdir(parents=True, exist_ok=True)
+
+            unchanged = music_dir / "unchanged.mp3"
+            removed = music_dir / "removed.mp3"
+            added = music_dir / "added.mp3"
+            touch_text(unchanged, "same")
+            touch_text(removed, "gone later")
+
+            add_tracks(
+                db,
+                [
+                    make_fs_track(unchanged, artist="Artist 1", album="Album 1", title="Unchanged"),
+                    make_fs_track(removed, artist="Artist 2", album="Album 2", title="Removed"),
+                ],
+            )
+            db.close()
+
+            removed.unlink()
+            touch_text(added, "new file")
+
+            calls: list[str] = []
+
+            def fake_new_fs_track(path: str, *, signature=None):
+                calls.append(Path(path).name)
+                if Path(path).name == "added.mp3":
+                    return FsTrack(
+                        file_path=path,
+                        file_name="added.mp3",
+                        title="Added",
+                        album="Album 3",
+                        artist="Artist 3",
+                        album_artist="Artist 3",
+                        duration=210.0,
+                        txt_lyrics=None,
+                        lrc_lyrics=None,
+                        track_number=1,
+                        modified_time=signature[0] if signature else None,
+                        file_size=signature[1] if signature else None,
+                    )
+                return None
+
+            scanner = LibraryScanner(db_path, [str(music_dir)])
+            with patch("ui.workers.library_scanner.new_fs_track_from_path", side_effect=fake_new_fs_track):
+                scanner.run()
+
+            db2 = sqlite3.connect(db_path)
+            db2.row_factory = sqlite3.Row
+            try:
+                index = get_library_file_index(db2)
+                names = {Path(path).name for path in index}
+                self.assertEqual(names, {"unchanged.mp3", "added.mp3"})
+                self.assertEqual(calls, ["added.mp3"])
+            finally:
+                db2.close()
